Modify rout for adding jne sending group
